+++ src/lib.rs
@@ -6,7 +6,7 @@ mod r#macro;
 mod render;
 
 pub use attrs::Attrs;
-pub use render::{Fn, Raw, Render};
+pub use render::{Raw, Render};
 
 /// the canonical html doctype
 pub const DOCTYPE: Raw<&str> = Raw("<!DOCTYPE html>");